requirements.txt
PyMuPDF
groq
ollama
openai
google-generativeai
fastapi
python-dotenv
uvicorn
websockets